Extract generic Drivers and OdeControl to ode_interface.cpp
- Created src/ode_interface.cpp with generic functions
- Drivers_* functions (new, set_constant, set_variable, evaluate, etc.)
- OdeControl_new() function
- Removed these functions from lorenz_interface.cpp
- These are system-agnostic and can be used by any model
